move Hyperlink to api and create HyperlinkWriter
