+++ src/api/draw/Image.ts
@@ -5,9 +5,9 @@ import { ImageWriter } from "../../xml/draw/ImageWriter";
 
 /**
  * This class represents an image in a paragraph.
- * 
+ *
  * It is used to embed image data in BASE64 encoding.
- * 
+ *
  * @example
  * document.getBody()
  *   .addParagraph()
@@ -22,7 +22,7 @@ export class Image extends OdfElement {
 
   /**
    * Creates an image
-   * 
+   *
    * @example
    * const image = new Image("/home/homer/myself.png");
    *
@@ -37,11 +37,11 @@ export class Image extends OdfElement {
 
   /**
    * The `getPath()` method returns the path to the image file that should be embedded.
-   * 
+   *
    * @example
    * const image = new Image("/home/homer/myself.png");
    * meta.getPath(); // '/home/homer/myself.png'
-   * 
+   *
    * @returns {string} The path to the image file
    * @since 0.7.0
    */
@@ -51,7 +51,7 @@ export class Image extends OdfElement {
 
   /**
    * Sets the new style of this image.
-   * 
+   *
    * @example
    * const image = new Image("/home/homer/myself.png");
    * meta.setStyle(new ImageStyle());
@@ -76,7 +76,7 @@ export class Image extends OdfElement {
    * image.getStyle();                // default style
    * meta.setStyle(new ImageStyle());
    * image.getStyle();                // previously set style
-   * 
+   *
    * @returns {IImageStyle} The style of the image
    * @since 0.5.0
    */
